clang warnings/errors

+++ trunk/linbox/linbox/solutions/solve.h
@@ -542,7 +542,7 @@ namespace LinBox
 
 		typedef Modular<double> Field;
 		// 0.7213475205 is an upper approximation of 1/(2log(2))
-		RandomPrimeIterator genprime( 26-(int)ceil(log((double)A.rowdim())*0.7213475205));
+		RandomPrimeIterator genprime((unsigned int)( 26-(int)ceil(log((double)A.rowdim())*0.7213475205)));
 		RationalSolver<Ring, Field, RandomPrimeIterator, DixonTraits> rsolve(A.field(), genprime);
 		SolverReturnStatus status = SS_OK;
 
@@ -639,7 +639,7 @@ namespace LinBox
 
 		typedef Modular<double> Field;
 		// 0.7213475205 is an upper approximation of 1/(2log(2))
-		RandomPrimeIterator genprime( 26-(int)ceil(log((double)A.rowdim())*0.7213475205));
+		RandomPrimeIterator genprime((unsigned int) (26-(int)ceil(log((double)A.rowdim())*0.7213475205)));
 		RationalSolver<Ring, Field, RandomPrimeIterator, SparseEliminationTraits> rsolve(A.field(), genprime);
 		SolverReturnStatus status = SS_OK;
 
@@ -853,7 +853,7 @@ namespace LinBox
 
 		commentator().start ("Integer CRA Solve", "Isolve");
 
-		RandomPrimeIterator genprime( 26 -(int)ceil(log((double)A.rowdim())*0.7213475205));
+		RandomPrimeIterator genprime((unsigned int)( 26 -(int)ceil(log((double)A.rowdim())*0.7213475205)));
 		//         RationalRemainder< Modular<double> > rra((double)
 		//                                                  ( A.coldim()/2.0*log((double) A.coldim()) ) );
 
@@ -927,11 +927,11 @@ namespace LinBox
 					  mpz_t mp_D ;
 					  mpz_init(mp_D);
 				  if (!m.computeRNS()) {
-					  IML::nonsingSolvLlhsMM(IML::RightSolu,B.rowdim(),1,
+					  IML::nonsingSolvLlhsMM(IML::RightSolu,(long)B.rowdim(),1,
 								 mp_A, mp_B, mp_N, mp_D);
 				  }
 				  else {
-					  long n = B.coldim();
+					  long n = (long)B.coldim();
 					  long basislen = 1;
 					  mpz_t mp_alpha, mp_maxInter;
 					  IML::FiniteField qh;
@@ -947,10 +947,10 @@ namespace LinBox
 					  mpz_clear(mp_maxInter);
 					  mpz_clear(mp_alpha);
 					  /*  CRNS[i] = [A_11, A_12] mod basis[i] */
-					  ARNS = IML_XMALLOC(IML::Double *, basislen);
+					  ARNS = IML_XMALLOC(IML::Double *, (size_t)basislen);
 					  for (long i = 0; i < basislen; ++i)
 					  {
-						  ARNS[i] = IML_XMALLOC(IML::Double, n*n);
+						  ARNS[i] = IML_XMALLOC(IML::Double,(size_t)(n*n));
 						  for (long j = 0; j < n; ++j)
 							  for (long l = 0; l < n; l++)
 								  ARNS[i][j*n+l] = (IML::Double) mpz_fdiv_ui(mp_A[j * n + l], basis[i]);
@@ -982,14 +982,14 @@ namespace LinBox
 				   }
 				   if ( !m.reduced() ) {
 					   IML::certSolveMP(m.certificate(),
-							    B.rowdim(),B.coldim(),
+							   (long) B.rowdim(),(long)B.coldim(),
 							    mp_A, mp_b, mp_N, mp_D,
 							    mp_NZ, mp_DZ);
 				   }
 				   else {
 					   IML::certSolveRedMP(m.certificate(),
 							       m.nullcol() //NULLSPACE_COLUMN
-							       ,B.rowdim(),B.coldim(),
+							       ,(long)B.rowdim(),(long)B.coldim(),
 							       mp_A, mp_b, mp_N, mp_D,
 							       mp_NZ, mp_DZ);
